Generate test election from the dev page
