+++ frontend/src/features/dev/components/DevHomePage.tsx
@@ -6,11 +6,14 @@ import { useApiState } from "@/api/useApiState";
 import { useInitialApiGet } from "@/api/useInitialApiGet";
 import { PageTitle } from "@/components/page_title/PageTitle";
 import { AppLayout } from "@/components/ui/AppLayout/AppLayout";
+import { Button } from "@/components/ui/Button/Button";
 import { Loader } from "@/components/ui/Loader/Loader";
+import { Modal } from "@/components/ui/Modal/Modal";
 import { useUserRole } from "@/hooks/user/useUserRole";
 import { t } from "@/i18n/translate";
 import { Election, ELECTION_LIST_REQUEST_PATH, ElectionListResponse, LoginResponse } from "@/types/generated/openapi";
 
+import { GenerateTestElectionForm } from "./GenerateTestElectionForm";
 import { MockTest } from "./MockTest";
 
 function Links() {
@@ -104,6 +107,7 @@ function AdministratorCoordinatorLinks({ electionList }: LinksProps) {
 function DevLinks() {
   const { user, login, logout } = useApiState();
   const [response, setResponse] = useState<ApiResult<LoginResponse> | null>(null);
+  const [showGenerateElectionModal, setShowGenerateElectionModal] = useState(false);
 
   if (response !== null && isError(response)) {
     throw response;
@@ -132,6 +136,18 @@ function DevLinks() {
             </a>
           </li>
         )}
+        <li>
+          <Button
+            type="button"
+            variant="underlined"
+            size="md"
+            onClick={() => {
+              setShowGenerateElectionModal(true);
+            }}
+          >
+            Genereer testverkiezing
+          </Button>
+        </li>
       </ul>
       {!__API_MSW__ && (
         <>
@@ -213,6 +229,18 @@ function DevLinks() {
         </>
       )}
       {(__API_MSW__ || user) && <Links />}
+      {!__API_MSW__ && showGenerateElectionModal && (
+        <Modal
+          title="Genereer testverkiezing"
+          onClose={() => {
+            setShowGenerateElectionModal(false);
+          }}
+          noFlex
+          autoWidth
+        >
+          <GenerateTestElectionForm />
+        </Modal>
+      )}
     </>
   );
 }
